Updated references in rendering pages.
